video_to_dots/requirements.txt
# Dot Matrix tools: image handling for convert.py
Pillow>=10.0.0
